Split ConfigurationButton into factory and button class. Had to fix PMD definition of naming convention
